+++ add.php
@@ -1,19 +1,9 @@
 <?php
-date_default_timezone_set('Europe/Moscow');
-
-require_once 'src/utils.php';
-$connection = setup_connection();
+require_once 'src/common.php';
 
 require_once 'src/user_queries.php';
 $current_user = get_random_user($connection)[0];
 $current_user['avatar'] = 'img/user.jpg';
-
-require_once 'src/category_queries.php';
-$categories = get_all_categories($connection);
-
-$navigation = include_template('templates/navigation', [
-    'categories' => $categories
-]);
 
 $lot = [];
 $errors = [];
