+++ lot.php
@@ -1,18 +1,5 @@
 <?php
-date_default_timezone_set('Europe/Moscow');
-
-$current_user = rand(0, 1) ? [
-    'id' => 3,
-    'name' => 'Женёк Пыхарь',
-    'avatar' => 'img/user.jpg'
-] : null;
-
-
-require_once 'src/utils.php';
-$connection = setup_connection();
-
-require_once 'src/category_queries.php';
-$categories = get_all_categories($connection);
+require_once 'src/common.php';
 
 require_once 'src/lot_queries.php';
 $lot = get_lot_by_id($connection, $_GET['id'] ?? null);
