+++ src/common.php
@@ -1,0 +1,10 @@
+<?php
+date_default_timezone_set('Europe/Moscow');
+
+$current_user = rand(0, 1) ? [
+    'name' => 'Женёк Пыхарь',
+    'avatar' => 'img/user.jpg'
+] : null;
+
+require_once 'db.php';
+require_once 'navigation.php';
