+++ templates/user_menu.php
@@ -8,7 +8,7 @@
 <?php else: ?>
     <ul class="user-menu__list">
         <li class="user-menu__item">
-            <a href="#">Регистрация</a>
+            <a href="register.php">Регистрация</a>
         </li>
         <li class="user-menu__item">
             <a href="#">Вход</a>
